Completed in November 2017
joyReader is a simple webapp developed for sharing book resources among
acquaintances at the end of last year (2017/11) due to participating in
a certain school competition. As an obscure coder, it is entirely
responsible for self-learning vue-related knowledge, maybe some code is
not very well written and some bugs. Welcome to make relevant
suggestions in issue` Thank you.